+++ packages/bot-api/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.45](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.44...@wireapp/bot-api@4.5.45) (2019-08-08)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.44](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.43...@wireapp/bot-api@4.5.44) (2019-08-08)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.43](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.42...@wireapp/bot-api@4.5.43) (2019-08-07)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.42](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.41...@wireapp/bot-api@4.5.42) (2019-08-07)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.41](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.40...@wireapp/bot-api@4.5.41) (2019-08-07)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.40](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.39...@wireapp/bot-api@4.5.40) (2019-08-06)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.39](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.38...@wireapp/bot-api@4.5.39) (2019-08-06)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [4.5.38](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.37...@wireapp/bot-api@4.5.38) (2019-08-05)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
 ## [4.5.37](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@4.5.36...@wireapp/bot-api@4.5.37) (2019-08-02)
 
 **Note:** Version bump only for package @wireapp/bot-api