+++ packages/react-ui-kit/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.11.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@7.11.0...@wireapp/react-ui-kit@7.11.1) (2019-08-07)
+
+
+### Bug Fixes
+
+* **react-ui-kit:** Keep background color on overscroll ([#2169](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/issues/2169)) ([e52f5f2](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/commit/e52f5f2))
+
+
+
+
+
+# [7.11.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@7.10.1...@wireapp/react-ui-kit@7.11.0) (2019-08-06)
+
+
+### Features
+
+* **react-ui-kit:** Themes ([#2158](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/issues/2158)) ([e5f5d69](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/commit/e5f5d69))
+
+
+
+
+
+## [7.10.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@7.10.0...@wireapp/react-ui-kit@7.10.1) (2019-08-05)
+
+**Note:** Version bump only for package @wireapp/react-ui-kit
+
+
+
+
+
 # [7.10.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@7.9.3...@wireapp/react-ui-kit@7.10.0) (2019-08-02)