+++ packages/store-engine-dexie/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.3](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.5.2...@wireapp/store-engine-dexie@0.5.3) (2019-08-08)
+
+**Note:** Version bump only for package @wireapp/store-engine-dexie
+
+
+
+
+
+## [0.5.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.5.1...@wireapp/store-engine-dexie@0.5.2) (2019-08-08)
+
+**Note:** Version bump only for package @wireapp/store-engine-dexie
+
+
+
+
+
+## [0.5.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.5.0...@wireapp/store-engine-dexie@0.5.1) (2019-08-07)
+
+**Note:** Version bump only for package @wireapp/store-engine-dexie
+
+
+
+
+
+# [0.5.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.4.7...@wireapp/store-engine-dexie@0.5.0) (2019-08-07)
+
+
+### Features
+
+* **store-engine,store-engine-bro-fs,store-engine-dexie,store-engine-fs,store-engine-sqleet:** Auto-increment primary keys ([#2163](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/issues/2163)) ([0d02747](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/commit/0d02747))
+
+
+
+
+
+## [0.4.7](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.4.6...@wireapp/store-engine-dexie@0.4.7) (2019-08-06)
+
+**Note:** Version bump only for package @wireapp/store-engine-dexie
+
+
+
+
+
+## [0.4.6](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.4.5...@wireapp/store-engine-dexie@0.4.6) (2019-08-05)
+
+**Note:** Version bump only for package @wireapp/store-engine-dexie
+
+
+
+
+
 ## [0.4.5](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-dexie/compare/@wireapp/store-engine-dexie@0.4.4...@wireapp/store-engine-dexie@0.4.5) (2019-08-01)
 
 **Note:** Version bump only for package @wireapp/store-engine-dexie