+++ packages/store-engine-fs/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.3](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.4.2...@wireapp/store-engine-fs@1.4.3) (2019-08-08)
+
+**Note:** Version bump only for package @wireapp/store-engine-fs
+
+
+
+
+
+## [1.4.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.4.1...@wireapp/store-engine-fs@1.4.2) (2019-08-08)
+
+**Note:** Version bump only for package @wireapp/store-engine-fs
+
+
+
+
+
+## [1.4.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.4.0...@wireapp/store-engine-fs@1.4.1) (2019-08-07)
+
+**Note:** Version bump only for package @wireapp/store-engine-fs
+
+
+
+
+
+# [1.4.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.3.7...@wireapp/store-engine-fs@1.4.0) (2019-08-07)
+
+
+### Features
+
+* **store-engine,store-engine-bro-fs,store-engine-dexie,store-engine-fs,store-engine-sqleet:** Auto-increment primary keys ([#2163](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/issues/2163)) ([0d02747](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/commit/0d02747))
+
+
+
+
+
+## [1.3.7](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.3.6...@wireapp/store-engine-fs@1.3.7) (2019-08-06)
+
+**Note:** Version bump only for package @wireapp/store-engine-fs
+
+
+
+
+
+## [1.3.6](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.3.5...@wireapp/store-engine-fs@1.3.6) (2019-08-05)
+
+**Note:** Version bump only for package @wireapp/store-engine-fs
+
+
+
+
+
 ## [1.3.5](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine-fs/compare/@wireapp/store-engine-fs@1.3.4...@wireapp/store-engine-fs@1.3.5) (2019-08-01)
 
 **Note:** Version bump only for package @wireapp/store-engine-fs